requirements.txt
opencv-python
mediapipe
pillow
numpy
torch
torchvision
streamlit
streamlit-webrtc
